+++ requirements.txt
@@ -1,4 +1,4 @@
-pytorch-lightning >= 0.9.0
+pytorch-lightning >= 1.0.0
 torch >= 1.10.0
 transformers >= 3.2.0
 kaggle >= 1.5.8
