+++ scripts/compute_bias_metric.py
@@ -3,7 +3,18 @@ import json
 import numpy as np
 from sklearn.metrics import roc_auc_score
 import argparse
-from utils import compute_auc, compute_subgroup_auc
+
+
+def compute_auc(y_true, y_pred):
+    try:
+        return roc_auc_score(y_true, y_pred)
+    except ValueError:
+        return np.nan
+
+
+def compute_subgroup_auc(df, subgroup, label, model_name):
+    subgroup_examples = df[df[subgroup]]
+    return compute_auc(subgroup_examples[label], subgroup_examples[model_name])
 
 
 def compute_bpsn_auc(df, subgroup, label, model_name):
