+++ scripts/compute_language_breakdown.py
@@ -1,7 +1,18 @@
 import pandas as pd
 import json
 import argparse
-from utils import compute_subgroup_auc
+
+
+def compute_auc(y_true, y_pred):
+    try:
+        return roc_auc_score(y_true, y_pred)
+    except ValueError:
+        return np.nan
+
+
+def compute_subgroup_auc(df, subgroup, label, model_name):
+    subgroup_examples = df[df[subgroup]]
+    return compute_auc(subgroup_examples[label], subgroup_examples[model_name])
 
 
 def convert_dataframe_to_bool(df):
