+++ scripts/convert_weights.py
