+++ scripts/train.py
@@ -8,8 +8,8 @@ from pytorch_lightning.callbacks import ModelCheckpoint
 from torch.nn import functional as F
 from torch.utils.data import DataLoader
 
-import src.data_loaders as module_data
-from src.utils import get_model_and_tokenizer
+import detoxify.utils.data_loaders as module_data
+from detoxify.utils.extra import get_model_and_tokenizer
 
 
 class ToxicClassifier(pl.LightningModule):
